Trying to solve issue with page fault on any address large than 1GB

+++ managers/filesystem/filesystem.c
@@ -111,7 +111,7 @@ Filesystem_SeekFile(UID fd,
 {
     if(!initialized)return -1;
     FileDescriptor *desc = (FileDescriptor*)Filesystem_FindDescriptorFromUID(fd);
-    uint8_t result = desc->driver->_H_Filesystem_SeekFile(desc, EXTRACT_ID(fd), offset, whence);
+    uint64_t result = desc->driver->_H_Filesystem_SeekFile(desc, EXTRACT_ID(fd), offset, whence);
     return result;
 }
